Move all the game logic to GameLogic from Prompter, ab#502, ab#504
