+++ src/main/java/com/zombiecastlerush/util/Inputs.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 
 class Inputs {
     public static String getUserInput(String displayMessage) {
-        System.out.printf(displayMessage + "\n>");
+        System.out.printf(displayMessage + "\n> ");
         Scanner sc = new Scanner(System.in);
         return sc.nextLine();
     }
